+++ examples/dataProvider/airtable/src/pages/posts/edit.tsx
@@ -26,8 +26,9 @@ export const PostEdit: React.FC<IResourceComponentsProps> = () => {
         defaultValue: postData?.category,
     });
 
-    const [selectedTab, setSelectedTab] =
-        useState<"write" | "preview">("write");
+    const [selectedTab, setSelectedTab] = useState<"write" | "preview">(
+        "write",
+    );
 
     return (
         <Edit saveButtonProps={saveButtonProps}>