+++ packages/live-previews/next.config.js
@@ -1,0 +1,8 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+    reactStrictMode: true,
+    swcMinify: true,
+    productionBrowserSourceMaps: true,
+};
+
+module.exports = nextConfig;